app/requirements.txt
pyyaml
requests
pyTelegramBotAPI